+++ CHANGELOG.md
@@ -1,1 +1,4 @@
 ### v1.0.0
+
+Database client created
+- pressing button sends standard text to database